Moved Availability to its own package
