+++ src/test/java/staffconnect/model/person/AvailabilityTest.java
@@ -8,6 +8,7 @@ import static staffconnect.testutil.Assert.assertThrows;
 import java.time.DayOfWeek;
 
 import org.junit.jupiter.api.Test;
+import staffconnect.model.Availability.Availability;
 
 public class AvailabilityTest {
